+++ uui/src/lib/shared/theme/config/theme-config.types.d.ts
@@ -5,3 +5,8 @@ export type CssProperties = Properties & PropertiesHyphen;
 export interface CssObject {
   [key: string]: CssProperties;
 }
+
+export type BaseScopedStyle = CssProperties | { [key: string]: CssProperties };
+export type ScopedStyles = { $self?: BaseScopedStyle } & {
+  [key: string]: BaseScopedStyle;
+} & CssProperties;